moving around Hole and Name
