+++ semantic-scope-graph/src/Data/Name.hs
@@ -1,5 +1,6 @@
-{-# LANGUAGE FlexibleContexts, OverloadedStrings #-}
-module Data.Abstract.Name
+{-# LANGUAGE FlexibleContexts  #-}
+{-# LANGUAGE OverloadedStrings #-}
+module Data.Name
 ( Name
 -- * Constructors
 , gensym
@@ -12,8 +13,9 @@ module Data.Abstract.Name
 import           Control.Effect.Fresh
 import           Data.Aeson
 import qualified Data.Char as Char
+import           Data.Hashable
+import           Data.Text (Text)
 import qualified Data.Text as Text
-import           Prologue
 
 -- | The type of variable names.
 data Name
